first phase of rip and replace.  Copied all "new" files over.
