+++ .gitignore
@@ -4,3 +4,4 @@
 /target/streams/
 /target/scala-2.12/classes/
 /target/scala-2.12/resolution-caches/
+/.idea/
