+++ .idea/libraries/sbt__org_msgpack_msgpack_core_0_8_21_jar.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="sbt: org.msgpack:msgpack-core:0.8.21:jar">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/Library/Caches/Coursier/v1/https/repo1.maven.org/maven2/org/msgpack/msgpack-core/0.8.21/msgpack-core-0.8.21.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$USER_HOME$/Library/Caches/Coursier/v1/https/repo1.maven.org/maven2/org/msgpack/msgpack-core/0.8.21/msgpack-core-0.8.21-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$USER_HOME$/Library/Caches/Coursier/v1/https/repo1.maven.org/maven2/org/msgpack/msgpack-core/0.8.21/msgpack-core-0.8.21-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
