--- .idea/libraries/sbt__org_scala_lang_modules_scala_collection_compat_2_13_2_1_3_jar.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="sbt: org.scala-lang.modules:scala-collection-compat_2.13:2.1.3:jar">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/Library/Caches/Coursier/v1/https/repo1.maven.org/maven2/org/scala-lang/modules/scala-collection-compat_2.13/2.1.3/scala-collection-compat_2.13-2.1.3.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$USER_HOME$/Library/Caches/Coursier/v1/https/repo1.maven.org/maven2/org/scala-lang/modules/scala-collection-compat_2.13/2.1.3/scala-collection-compat_2.13-2.1.3-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$USER_HOME$/Library/Caches/Coursier/v1/https/repo1.maven.org/maven2/org/scala-lang/modules/scala-collection-compat_2.13/2.1.3/scala-collection-compat_2.13-2.1.3-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
