+++ .idea/sbt.xml
@@ -10,7 +10,7 @@
             <option value="$PROJECT_DIR$/project" />
           </set>
         </option>
-        <option name="sbtVersion" value="1.3.8" />
+        <option name="sbtVersion" value="1.3.13" />
       </SbtProjectSettings>
     </option>
   </component>
